Disabled the Insert key in all TextBox controls
The Insert key would toggle overtype mode in TextBox controls, but there is absolutely no visual feedback. Pressing the key accidentally, any future input will overwrite existing text. This may lead to data loss if the user isn't aware of it and doesn't pay much attention to it. So it's safer to completely disable that key (without modifiers).