Changed tool bar button styles to Windows 8 style, for all platforms
The borders and backgrounds were darker on Windows 7 but the borders were completely missing on Windows 8, so I made a local copy of the ToolBar styles with Blend on Windows 7 and updated the colours from Windows 8.1 screenshots. Now the behaviour is closer to Windows, just the colours are a bit lighter for Windows 7.