Add LoadFromStream() void
Move most of the body of LoadFromEmbeddedResource() into a new void which just accepts a stream and loads the definitions. LoadFromEmbeddedResource() calls LoadFromStream() after the resource stream was received. This adds another abstract layer, all data can be represented as a stream so it will be possible to load the translation through a networkstream, a http response stream, a custom resource stream, etc.